reorganized common models
