+++ components/Common Models/Modals/CustomModal.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Modal, View, Text, StyleSheet } from 'react-native';
-import { CommonButton } from './CommonButton';
+import { CommonButton } from '../Buttons/CommonButton';
 
 // Custom modal for use in creating instructions
 // Allows for a custom header and body
